Fixed narrowing covnersion

+++ Bulb/src/Rendering/Camera.cpp
@@ -15,7 +15,7 @@ namespace blb::rnd {
 
 	Camera::Camera(clv::plt::Window& window, const ProjectionMode projection) {
 		window.onWindowResize.bind([this](const mth::vec2ui& size) {
-			setViewport({ 0, 0, size.x, size.y });
+			setViewport({ 0, 0, static_cast<int32_t>(size.x), static_cast<int32_t>(size.y) });
 		});
 
 		renderTarget = window.getSurface()->getRenderTarget();
